requirements.txt
Django==5.0.4
Pillow>=11.0.0

setuptools~=65.5.0
asgiref~=3.8.1
sqlparse~=0.5.3
psycopg2~=2.9.9
requests~=2.31.0
django-nested-admin~=4.1.4

gunicorn
psycopg2-binary
dj-database-url
whitenoise